+++ src/snakemake/exceptions.py
@@ -384,7 +384,7 @@ class AmbiguousRuleException(RuleException):
         wildcards_a = utils.format("{}", job_a._format_wildcards)
         wildcards_b = utils.format("{}", job_b._format_wildcards)
         super().__init__(
-            "Rules {job_a} and {job_b} are ambiguous for the file {f}.\n"
+            "Rules {job_a} and {job_b} are ambiguous for the file {fmt_iofile(f)}.\n"
             "Consider starting rule output with a unique prefix, constrain "
             "your wildcards, or use the ruleorder directive.\n"
             "Wildcards:\n"
